# pkg/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Streaming service YouTube alike",
        "title": "ROSS API",
        "contact": {
            "name": "Ro Ngoc Vo",
            "url": "github.com/rovn208",
            "email": "ngocro208@gmail.com"
        },
        "version": "0.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/follows/followers": {
            "get": {
                "description": "Get list follower",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "follows"
                ],
                "summary": "Get list follower",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "20",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "0",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.listFollowResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "{\"error\": \"error message\"}",
                        "schema": {}
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "follows"
                ],
                "summary": "Follow user",
                "parameters": [
                    {
                        "description": "123456789",
                        "name": "following_user_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"messsage\": \"follow user successfully\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"error\": \"error message\"}",
                        "schema": {}
                    },
                    "500": {
                        "description": "{\"error\": \"error message\"}",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "follows"
                ],
                "summary": "Unfollow user",
                "parameters": [
                    {
                        "description": "123456789",
                        "name": "following_user_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"messsage\": \"unfollow user successfully\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"error\": \"error message\"}",
                        "schema": {}
                    },
                    "500": {
                        "description": "{\"error\": \"error message\"}",
                        "schema": {}
                    }
                }
            }
        },
        "/api/v1/follows/following": {
            "get": {
                "description": "Get list following",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "follows"
                ],
                "summary": "Get list following",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "20",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "0",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.listFollowResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "{\"error\": \"error message\"}",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.listFollowResponse": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}